Resolve cyclic include spack.cmd <-> spack.extensions.
